+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Log
 
+## [0.0.5] - 2020-11-09
+
+- Implemented *Go to definition* for methods using `F12`
+
 ## [0.0.4] - 2020-11-06
 
 - Implemented *Go to symbol* using `CTRL+Shift+O` or `CMD+Shift+O` 
 
 ## [0.0.3] - 2020-10-22
 
-- Implemented *Go to definition* using `F12`
+- Implemented *Go to definition* for include files using `F12`
 
 ## [0.0.2] - 2020-10-21